moved logs writer implementations into separate package.  splited mongo logs writing into separate collections
